+++ CHANGELOG.md
@@ -23,7 +23,7 @@ following procedure to upgrade all the nodes running a head:
 Only when this procedure has been applied to all Hydra nodes can you open a new
 head again.
 
-- **BREAKING** Hydra TUI now uses `--cardanoSigningKey` flag to specify a user
+- **BREAKING** Hydra TUI now uses `--cardano-signing-key` flag to specify a user
   key used to select UTxO and submit a commit transaction.
 
 - Fuel removal: